requirements.txt
# python 3.10.0

prefect==2.14.17
prefect-dbt[bigquery]==0.4.1
prefect-airbyte==0.3.2
Flask==3.0.2
waitress==3.0.0
python-dotenv==1.0.1